int -> long

+++ bitstamp/src/main/java/net/parasec/trading/ticker/bitstamp/BitstampMarshalling.java
@@ -22,7 +22,7 @@ public final class BitstampMarshalling {
     final int cents = (int) Math.round(rawOrder.getPrice()*100);
     final long satoshi = (long) Math.round(rawOrder.getAmount()*100000000);
     final OrderInfo orderInfo
-        = new OrderInfo(Integer.toString(rawOrder.getId()), cents, satoshi, 
+        = new OrderInfo(Long.toString(rawOrder.getId()), cents, satoshi, 
               rawOrder.getDatetime()); 
     return new OrderEvent(state, direction, "btc/usd", "bitstamp", orderInfo); 
   }
@@ -55,4 +55,3 @@ public final class BitstampMarshalling {
     return new Trade(null, cents, satoshi, 0, 0, rawTrade.getId(), null, null);
   }
 }
-
